Adding a number of files in as I experimented a bit. Changes aren't very granular but it should now be easy to add and delete projects and people with Markdown (See the test examples in _projects and _people).
In _data, faculty.yml and phd.yml were added. staff.yml and alumni.yml will also need
to be added in a similar fashion.

I'm not certain how sponsors should be displayed, so I just threw them on a page
and left them as is. Sponsors are also displayed at the footer of the index page.

Feedback is welcome :)

TODO:
1. clean up "events" - maybe merge this category with "news"?
2. publications
3. port over existing data on systems.soe.ucsc.edu
4. ???