Added "edit me" links on every page. If you want to provide an offsite link to your page, just add "offsite: true" to your entry in the corresponding YAML page, since you wouldn't have a page to edit.
The links are to the master branch's pages, so because the experiment
branch hasn't yet been merged with master, it might link you to a 404 page on
github.

If you edit sponsors, be sure to include an image of the logo as well in /images/Sponsors/.

We can think of a way to make these edit links look better later. :)